requirements-dev.txt
# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting & Formatting
flake8>=6.1.0
black>=23.7.0
pylint>=2.17.0
mypy>=1.5.0

# Type stubs
types-openpyxl>=3.1.0
pandas-stubs>=2.0.0
